fix: rename labyrinth
