+++ autopcr/module/modules/labyrinth.py
